Suporte a meio-espaço (0,5) na carga
- espacos vira Double em ItemJogo e ItemPersonagem; cargaAtual no response vira double
- Alquímicos ocupam 0,5 espaço (como no livro)
- Sobrecarga: penalidade de Agilidade usa o piso do excesso (só ponto inteiro conta)
- Frontend: inputs de espaço com step 0.5; carga/espaços formatados em pt-BR (0,5)

+++ asus-platform/src/main/java/com/asus/platform/web/dto/PersonagemResponse.java
@@ -26,7 +26,7 @@ public record PersonagemResponse(
         StatusDto status,
         int deslocamento,
         int cargaMaxima,
-        int cargaAtual,
+        double cargaAtual,
         int limiteHabilidades,
         int limiteFeiticos,
         int limiteBencaos,
